bitbar-devicepool errata: round 2 (#28)

+++ modules/roles_profiles/manifests/profiles/remove_bootstrap_user.pp
@@ -6,9 +6,11 @@ class roles_profiles::profiles::remove_bootstrap_user {
 
     $user = 'relops'
 
-    exec { "${user}_admin_group":
-        command => "/usr/bin/dscl . -delete /Users/${user} || rm /private/var/db/dslocal/nodes/Default/users/${user}.plist",
-        onlyif  => "/usr/bin/id ${user}",
+    if $::operatingsystem == 'Darwin' {
+        exec { "${user}_admin_group":
+            command => "/usr/bin/dscl . -delete /Users/${user} || rm /private/var/db/dslocal/nodes/Default/users/${user}.plist",
+            onlyif  => "/usr/bin/id ${user}",
+        }
     }
 
     user { $user:
